refactor: OAuth2 리다이렉트 URL 결정 로직을 공통 유틸리티로 추출
OAuth2AuthenticationSuccessHandler와 OAuth2AuthenticationFailureHandler에서
중복된 리다이렉트 URL 결정 로직을 OAuth2RedirectUrlResolver로 분리하여
코드 중복을 제거하고 유지보수성을 향상시켰습니다.

주요 변경사항:
- OAuth2RedirectUrlResolver: 리다이렉트 URL 결정 공통 유틸리티 클래스 추가
- SuccessHandler/FailureHandler: 중복된 getBaseUrl(), isLocalhost() 메서드 제거
- 각 핸들러는 OAuth2RedirectUrlResolver를 주입받아 사용하도록 변경
- coverage.gradle.kts: OAuth2RedirectUrlResolver를 커버리지 제외 대상에 추가
