+++ buildSrc/src/main/kotlin/plugin/coverage.gradle.kts
@@ -32,9 +32,10 @@ val excludeClassPatterns =
         "*.*ErrorCode*",
         "*.*Handler*",
 
-        // OAuth2 관련 일부 테스트는 E2E 테스트로 검증
+        // OAuth2 관련 일부 테스트는 E2E 테스트로 검증, 추후 개발이 어느정도 진행되면 삭제
         "*.CustomOAuth2UserService",
-        "*.CustomOAuth2AuthorizationRequestRepository"
+        "*.CustomOAuth2AuthorizationRequestRepository",
+        "*.OAuth2RedirectUrlResolver"
     )
 
 fun JacocoReportBase.configureClassDirectories() {
